environments/mcp_sandbox_env/pyproject.toml
[project]
name = "mcp-sandbox-env"
description = "Your environment description here"
tags = ["placeholder-tag", "train", "eval"]
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.2.0",
    "mcp>=1.16.0",
    "python-dotenv>=1.1.1",
    "verifiers",
]

[tool.uv.sources]
verifiers = { path = "./verifiers", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
